[PALANTIR ONLY] Make FROM=AUTHOR for revision emails
This is totally wrong, but it matches the behavior of our previous email
solution and we don't want to force people to update their email filters
when we switch from gitwatch to git-multimail.  There are lots of people
with lots of said Outlook filters (plus Outlook stupidly makes filtering
on Reply-To harder than filtering on From.  Outlook is annoyingly dumb.)

Signed-off-by: Elijah Newren <newren@palantir.com>